refactor: Modify project layout.
