+++ go/gildedrose/item/item.go
@@ -1,0 +1,6 @@
+package item
+
+type Item struct {
+	Name            string
+	SellIn, Quality int
+}
