+++ go/gildedrose/strategy/normal.go
@@ -1,7 +1,9 @@
-package gildedrose
+package strategy
+
+import "github.com/emilybache/gildedrose-refactoring-kata/gildedrose/item"
 
 type Normal struct {
-	*Item
+	*item.Item
 }
 
 func (item *Normal) UpdateSellIn() {
