+++ go/gildedrose/strategy/sulfuras.go
@@ -1,7 +1,9 @@
-package gildedrose
+package strategy
+
+import "github.com/emilybache/gildedrose-refactoring-kata/gildedrose/item"
 
 type Sulfuras struct {
-	*Item
+	*item.Item
 }
 
 func (item *Sulfuras) UpdateSellIn() {
